@@ -10,33 +10,34 @@
   "private": true,
   "dependencies": {
     "@angular-slider/ngx-slider": "^2.0.4",
-    "@angular/animations": "~13.3.11",
-    "@angular/cdk": "~12.2.4",
-    "@angular/common": "~13.3.11",
-    "@angular/compiler": "~13.3.11",
-    "@angular/core": "~13.3.11",
+    "@angular/animations": "^14.1.0",
+    "@angular/cdk": "^12.2.13",
+    "@angular/common": "^14.1.0",
+    "@angular/compiler": "^14.1.0",
+    "@angular/core": "^14.1.0",
     "@angular/flex-layout": "^12.0.0-beta.34",
-    "@angular/forms": "~13.3.11",
-    "@angular/material": "^12.2.4",
-    "@angular/platform-browser": "~13.3.11",
-    "@angular/platform-browser-dynamic": "~13.3.11",
-    "@angular/router": "~13.3.11",
+    "@angular/forms": "^14.1.0",
+    "@angular/material": "^12.2.13",
+    "@angular/platform-browser": "^14.1.0",
+    "@angular/platform-browser-dynamic": "^14.1.0",
+    "@angular/router": "^14.1.0",
+    "chart.js": "^3.8.2",
     "moment": "^2.29.3",
     "rxjs": "^6.6.6",
     "tslib": "^2.0.0",
     "zone.js": "^0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.1303.9",
-    "@angular-devkit/build-angular": "~13.3.9",
+    "@angular-devkit/architect": "^0.1401.0",
+    "@angular-devkit/build-angular": "^14.1.0",
     "@angular-eslint/builder": "12.1.0",
     "@angular-eslint/eslint-plugin": "12.1.0",
     "@angular-eslint/eslint-plugin-template": "12.1.0",
     "@angular-eslint/schematics": "12.5.0",
     "@angular-eslint/template-parser": "12.1.0",
-    "@angular/cli": "~13.3.9",
-    "@angular/compiler-cli": "~13.3.11",
-    "@angular/language-service": "^13.3.11",
+    "@angular/cli": "^14.1.0",
+    "@angular/compiler-cli": "^14.1.0",
+    "@angular/language-service": "^14.1.0",
     "@types/node": "^12.20.23",
     "@typescript-eslint/eslint-plugin": "4.23.0",
     "@typescript-eslint/parser": "4.23.0",
