@@ -1,4 +1,4 @@
-import { FormArray, FormGroup } from '@angular/forms';
+import { UntypedFormArray, UntypedFormGroup } from '@angular/forms';
 import { IModule } from '../api/commands.api';
 
 const enum ModuleFCN {
@@ -6,30 +6,30 @@ const enum ModuleFCN {
   cmds = 'commands'
 }
 
-export class ModuleCtrl extends FormGroup {
+export class ModuleCtrl extends UntypedFormGroup {
 
   name: string
 
   constructor(imodule: IModule) {
     super({})
     this.name = imodule.name
-    this.addControl(ModuleFCN.vars, new FormArray([]))
-    this.addControl(ModuleFCN.cmds, new FormArray([]))
+    this.addControl(ModuleFCN.vars, new UntypedFormArray([]))
+    this.addControl(ModuleFCN.cmds, new UntypedFormArray([]))
   }
 
   get varsFA() {
-    return this.get(ModuleFCN.vars) as FormArray;
+    return this.get(ModuleFCN.vars) as UntypedFormArray;
   }
 
-  set varsFA(fa: FormArray) {
+  set varsFA(fa: UntypedFormArray) {
     this.setControl(ModuleFCN.vars, fa);
   }
 
   get cmdsFA() {
-    return this.get(ModuleFCN.cmds) as FormArray;
+    return this.get(ModuleFCN.cmds) as UntypedFormArray;
   }
 
-  set cmdsFA(fa: FormArray) {
+  set cmdsFA(fa: UntypedFormArray) {
     this.setControl(ModuleFCN.cmds, fa);
   }
 
